hahooooooowa

+++ src/main/java/pfa/emsi/model/Admin.java
@@ -15,7 +15,7 @@ public class Admin implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idUser;
+	private Long idUser;
 
 	private String adresse;
 
@@ -50,11 +50,11 @@ public class Admin implements Serializable {
 		this.prenom = prenom;
 	}
 
-	public int getIdUser() {
+	public Long getIdUser() {
 		return this.idUser;
 	}
 
-	public void setIdUser(int idUser) {
+	public void setIdUser(Long idUser) {
 		this.idUser = idUser;
 	}
