+++ src/main/java/pfa/emsi/model/Directeur.java
@@ -16,7 +16,7 @@ public class Directeur implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idUser;
+	private Long idUser;
 
 	private String adresse;
 
@@ -59,11 +59,11 @@ public class Directeur implements Serializable {
 		this.prenom = prenom;
 	}
 
-	public int getIdUser() {
+	public Long getIdUser() {
 		return this.idUser;
 	}
 
-	public void setIdUser(int idUser) {
+	public void setIdUser(Long idUser) {
 		this.idUser = idUser;
 	}
